feat: 支持转换png jpeg

+++ main.h
@@ -4,9 +4,6 @@
 #include <string.h>
 #include <stdio.h>
 #include <sys/stat.h>
-#include "png.h"
-#include "jpeglib.h"
-#include "jerror.h"
 
 class Image {
   private:
@@ -15,7 +12,7 @@ class Image {
     uint32_t height;
     uint8_t **data;  // 解码后的位图数据
     uint8_t *buffer; // 编码后的文件数据
-    uint8_t bufferSize;
+    unsigned long bufferSize;
     Image();
     ~Image();
     void mallocImageData();
